更新作业和页面资源的处理

+++ grails-app/controllers/com/sanwn/mousika/FileResourceController.groovy
@@ -24,7 +24,11 @@ class FileResourceController {
 
     private def getFileRepo(String courseToken) {
         def path = request.servletContext.getRealPath(".")
-        return new File(path, "courseFiles/${courseToken}")
+        def fileRepo = new File(path, "courseFiles/${courseToken}/contents")
+        if (!fileRepo.exists()) {
+            FileUtils.forceMkdir(fileRepo)
+        }
+        return fileRepo
     }
 
     def upload() {
@@ -133,15 +137,17 @@ class FileResourceController {
 
         fileResourceInstance.properties = params
         def uploadedFile = request.getFile('qqfile')
-        def fileRepo = getFileRepo(Course.get(params.courseId).courseToken)
-        def filename = uploadedFile.originalFilename
-        def updatedFile = new File(fileRepo, filename)
-        uploadedFile.transferTo(updatedFile)
-        def oldFile = new File(fileResourceInstance.filePath)
-        oldFile.delete()
-        fileResourceInstance.filePath = updatedFile.getCanonicalPath()
-        def pos = filename.lastIndexOf('.')
-        fileResourceInstance.fileType = pos == -1 ? "*" : filename.substring(pos + 1)
+        def filename = uploadedFile.originalFilename.trim()
+        if (filename) {
+            def fileRepo = getFileRepo(Course.get(params.courseId).courseToken)
+            def updatedFile = new File(fileRepo, filename)
+            uploadedFile.transferTo(updatedFile)
+            def oldFile = new File(fileResourceInstance.filePath)
+            oldFile.delete()
+            fileResourceInstance.filePath = updatedFile.getCanonicalPath()
+            def pos = filename.lastIndexOf('.')
+            fileResourceInstance.fileType = pos == -1 ? "*" : filename.substring(pos + 1)
+        }
 
         if (!fileResourceInstance.save(flush: true)) {
             render(view: "edit", model: [fileResourceInstance: fileResourceInstance, course: Course.get(params.courseId)])
@@ -149,7 +155,7 @@ class FileResourceController {
         }
 
         flash.message = message(code: 'default.updated.message', args: [message(code: 'fileResource.label', default: 'FileResource'), fileResourceInstance.id])
-        redirect(action: "show", id: fileResourceInstance.id)
+        redirect(controller: 'course', action: "show", id: params.courseId)
     }
 
     def delete(Long id) {
